The scope of the variable ``plugins'' in slimv.vim is limited to s:.
As I started to used dein.vim, it caused a conflict on the variable
nambed ``plugins''; A type mistmatch on this variable was observed. I
roughly checked the use of the variable in slimv and found that script
local varible is proper for this variable. I do not want to check
dein.vim side because I have never read dein.vim.