[Bugfix] Swank boot commands in the case of screen is fixed.
SlimvSwankLoader function quotes both of `g:slimv_lisp` and the argument
of --load argument (the option name in SBCL), but it did not work as it
is expected in the case of screen. This commit fixes SlimvSwankCommand
and ensures to properly quote both of `g:slimv_lisp` and the --load
argument by adopting single quotes to quote each screen command.

If, for example, the path of `slime` directory included a whitespace,
then what SlimvSwankLoader returned was something like:

    "/usr/bin/sbcl" --load "/home/foo/personal swankdir/slime/start-swank.lisp"

Here, it is assumed that `let g:slimv_impl = 'sbcl'` and `let
g:slimv_lisp='/usr/bin/sbcl'` are written in .vimrc.
Finally, SlimvSwankCommand function executes following command if vim is
run on screen.

    ! screen -X eval "title slimv" "screen "/usr/bin/sbcl" --load "/home/foo/personal swankdir/slime/start-swank.lisp"" "select slimv"

For clarity, the commands exected by `screen -X eval` was as follows.

1. title slimv
2. screen /usr/bin/sbcl --load /home/foo/personal
3. swankdir/slime/start-swank.lisp
4. select slimv

Be aware that continuing 2 strings would not be recognized as 2
separated commands if they are not separated by whitespaces. This rule
holds even if one of such strings is not quoted.

But the commands we truly want should be as follows.

1. title slimv
2. screen "/usr/bin/sbcl" --load "/home/foo/personal swankdir/slime/start-swank.lisp"
3. select slimv

This commit fixes the string returned by SlimvSwankLoader in this
situation. The fixed returned value is as follows.

    ! screen -X eval 'title slimv' 'screen "/usr/bin/sbcl" --load "/home/foo/personal swankdir/slime/start-swank.lisp"' 'select slimv'