A trial implementation to keep multiple swank servers.
In this implementation, users are asked the swank server to connect on
every call of SlimvConnectServer. It means that users are asked on every
input of `<Leader>c`.

The swank server is specified by the port number. It is not satisfactory
from the conceptual view point because a port is not a server. But this
is quite easy to implement, works pretty fine as far as we use slimv on
localhost.

The interactive input line is filled by the value of `g:swank_port` at
that moment. The value of `g:swank_port` set in `.vimrc` works therefore
as the initial default port. From the second time of `<Leader>c`, the swank
server port in use at that moment  works as the default.

Each screen window that runs swank server is given a title composed of
the name of implementation and the port number. It may be useful when
users change the target swank server.

Next step ideas:

+ give a name to the each server
    - specify the target server by that name
+ give an explanatory text of each server